Inline code snippets for exception ruleset
