+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/exceptions/TooGenericExceptionSpec.kt
@@ -3,9 +3,8 @@ package io.gitlab.arturbosch.detekt.rules.exceptions
 import io.gitlab.arturbosch.detekt.api.internal.YamlConfig
 import io.gitlab.arturbosch.detekt.core.rules.createRuleSet
 import io.gitlab.arturbosch.detekt.core.rules.visitFile
-import io.gitlab.arturbosch.detekt.rules.Case
 import io.gitlab.arturbosch.detekt.rules.providers.ExceptionsProvider
-import io.gitlab.arturbosch.detekt.test.compileForTest
+import io.gitlab.arturbosch.detekt.test.compileContentForTest
 import io.gitlab.arturbosch.detekt.test.resource
 import org.assertj.core.api.Assertions.assertThat
 import org.spekframework.spek2.Spek
@@ -18,7 +17,7 @@ class TooGenericExceptionSpec : Spek({
         it("should not report any as all catch exception rules are deactivated") {
             val config = YamlConfig.loadResource(resource("deactivated-exceptions.yml").toURL())
             val ruleSet = ExceptionsProvider().createRuleSet(config)
-            val file = compileForTest(Case.TooGenericExceptions.path())
+            val file = compileContentForTest(tooGenericExceptionCode)
 
             val findings = ruleSet.visitFile(file)
 
@@ -26,3 +25,23 @@ class TooGenericExceptionSpec : Spek({
         }
     }
 })
+
+const val tooGenericExceptionCode = """
+        fun main() {
+        try {
+            throw Throwable()
+        } catch (e: ArrayIndexOutOfBoundsException) {
+            throw Error()
+        } catch (e: Error) {
+            throw Exception()
+        } catch (e: Exception) {
+        } catch (e: IllegalMonitorStateException) {
+        } catch (e: IndexOutOfBoundsException) {
+            throw RuntimeException()
+        } catch (e: Throwable) {
+        } catch (e: RuntimeException) {
+            throw NullPointerException()
+        } catch (e: NullPointerException) {
+
+        }
+    }"""
