+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/exceptions/TooGenericExceptionThrownSpec.kt
@@ -1,9 +1,8 @@
 package io.gitlab.arturbosch.detekt.rules.exceptions
 
 import io.gitlab.arturbosch.detekt.api.Config
-import io.gitlab.arturbosch.detekt.rules.Case
 import io.gitlab.arturbosch.detekt.test.TestConfig
-import io.gitlab.arturbosch.detekt.test.lint
+import io.gitlab.arturbosch.detekt.test.compileAndLint
 import org.assertj.core.api.Assertions.assertThat
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.describe
@@ -15,7 +14,7 @@ class TooGenericExceptionThrownSpec : Spek({
         it("should report one for each generic throw rules") {
             val rule = TooGenericExceptionThrown(Config.empty)
 
-            val findings = rule.lint(Case.TooGenericExceptions.path())
+            val findings = rule.compileAndLint(tooGenericExceptionCode)
 
             assertThat(findings).hasSize(thrownExceptionDefaults.size)
         }
@@ -24,7 +23,7 @@ class TooGenericExceptionThrownSpec : Spek({
             val config = TestConfig(mapOf(TooGenericExceptionThrown.THROWN_EXCEPTIONS_PROPERTY to "[MyException]"))
             val rule = TooGenericExceptionCaught(config)
 
-            val findings = rule.lint(Case.TooGenericExceptions.path())
+            val findings = rule.compileAndLint(tooGenericExceptionCode)
 
             assertThat(findings).isEmpty()
         }
